requirements.txt
pycryptodomex==3.15.0